Library

Types

Dimensionless.DimBasis — Type

DimBasis(basis_vectors...) -> DimBasis

Create a dimensional basis for a number of basis_vectors (quantities, units or dimensions). A string identifier can optionally be added to each basis vector.

Functions

Dimensionless.dim_matrix — Function

dim_matrix(basis_dims, all_values...)

Return the dimensional matrix for a set of basis dimensions basis_dims and all_values, a set of quantities, units or dimensions.

Dimensionless.number_of_dimensions — Function

number_of_dimensions(all_vars...)

Return the number of unique dimensions for the given variables all_vars.

Dimensionless.number_of_dimensionless — Function

number_of_dimensionless(all_vars...)

Return the number of dimensionless numbers that can be constructed for a problem characterized by the variables all_vars.

Dimensionless.dimensionless — Function

dimensionless(quantity, basis)

Make a quantity dimensionless using a dimensional basis.

Dimensionless.dimensionful — Function

dimensionful(value, unit, basis)

Restore the units of a dimensionless value using a dimensional basis.

Dimensionless.change_basis — Function

change_basis(var, basis, new_basis)

Transform the specified quantity or unit var from a current basis to a new_basis.

Dimensionless.print_dimensionless — Function

print_dimensionless([io, ]named_dim_vector, basis)

Print the dimensionless number that can be constructed using named_dim_vector in the specified dimensional basis.